Fix slider implementation for `ticksPerFrame`
The slider previously did set the value for `ticksPerFrame` correctly,
but now the `renderLoop` does not have to be called `ticksPerFrame`
times before the next frame is rendered on the screen. Instead, the
`renderLoop` is called once and the universe advances by `ticksPerFrame`
ticks and then renders the cells once, which is the intended behavior.