Redesign `live_neighbor_count` method
The method turned out to have a major impact on the overall app
performance, due to the utilization of the modulo operation. To improve
performance, the method was redesigned to rely on a `if` control flow
pattern to count the living neighbors of a given cell and handle the
special cases at the edge of the universe.