Fix invalid array length error
Before the fix, by defining a new `Uint8Array` over and over via
changing the `$gridSize` value, the above error would be thrown on
several occasions. By referencing the actual `$gridSize` store value
instead of the `width` and `height` value, the error is fixed. This is
most likely due to the speed of updates for each of the values when the
slider and handler method are triggered. The getter functions of the
canvas are seemingly slower than the update cycle of the slider and
handler method, causing a mismatch in the array length.